All application now use np.float64 values

+++ examples/fdtd-1d/autoparallel/fdtd-1d.py
@@ -37,7 +37,7 @@ def create_matrix(n_size):
 @constraint(ComputingUnits="${ComputingUnits}")
 @task(returns=1)
 def create_entry(index, n_size):
-    return np.float(np.float(index) / np.float(n_size))
+    return np.float64(np.float64(index) / np.float64(n_size))
 
 
 ############################################
@@ -145,8 +145,8 @@ if __name__ == "__main__":
     args = sys.argv[1:]
     NSIZE = int(args[0])
     TSIZE = int(args[1])
-    COEF1 = np.float(0.5)
-    COEF2 = np.float(0.7)
+    COEF1 = np.float64(0.5)
+    COEF2 = np.float64(0.7)
 
     # Log arguments if required
     if __debug__:
